created driver factory and fixed some touch actions bugs
. this is the first commit in the direction of moving towards a more
abstract driver factory which will support browsers, mobile, and desktop
applications
. BrowserFactory will now be used only for browser interactions and will
have a few number of methods.
. DriverFactory will now be used to initialize other driver types such
as SikuliX and Appium-based drivers
. All the changes were internal and should have no effects on current
users.
