+++ src/test/java/testPackage01/Test_SelectedValue.java
@@ -2,6 +2,7 @@ package testPackage01;
 
 import com.shaft.gui.browser.BrowserActions;
 import com.shaft.gui.browser.BrowserFactory;
+import com.shaft.gui.driver.DriverFactory.DriverType;
 import com.shaft.gui.element.ElementActions;
 import com.shaft.tools.io.ReportManager;
 import com.shaft.validation.Assertions;
@@ -12,7 +13,7 @@ import org.testng.annotations.Test;
 public class Test_SelectedValue {
     @Test
     public void simpleSelect() {
-        WebDriver driver = BrowserFactory.getBrowser(BrowserFactory.BrowserType.GOOGLE_CHROME);
+        WebDriver driver = BrowserFactory.getBrowser(DriverType.DESKTOP_CHROME);
         BrowserActions.navigateToURL(driver, "https://mdn.mozillademos.org/en-US/docs/Web/HTML/Element/select$samples/Basic_select?revision=1612304");
         By select = By.tagName("select");
         ElementActions.select(driver, select, "Third Value");
@@ -23,7 +24,7 @@ public class Test_SelectedValue {
 
     @Test
     public void multipleSelect() {
-        WebDriver driver = BrowserFactory.getBrowser(BrowserFactory.BrowserType.GOOGLE_CHROME);
+        WebDriver driver = BrowserFactory.getBrowser(DriverType.DESKTOP_CHROME);
         BrowserActions.navigateToURL(driver, "https://mdn.mozillademos.org/en-US/docs/Web/HTML/Element/select$samples/Advanced_select_with_multiple_features?revision=1612304");
         By select = By.tagName("select");
         ElementActions.select(driver, select, "Dog");
