+++ src/test/java/testPackage01/Test_sikulix.java
@@ -3,6 +3,7 @@ package testPackage01;
 import com.shaft.cli.TerminalActions;
 import com.shaft.gui.browser.BrowserActions;
 import com.shaft.gui.browser.BrowserFactory;
+import com.shaft.gui.driver.DriverFactory;
 import com.shaft.gui.element.ElementActions;
 import com.shaft.validation.Assertions;
 import org.openqa.selenium.WebDriver;
@@ -47,12 +48,12 @@ public class Test_sikulix {
     //@BeforeClass
     public void openApplication() {
         new TerminalActions().performTerminalCommand("calc");
-        calculator = BrowserFactory.getSikuliApp("Calculator");
+        calculator = DriverFactory.getSikuliApp("Calculator");
     }
 
     //@AfterClass
     public void closeApplication() {
-        BrowserFactory.closeSikuliApp(calculator);
+    	DriverFactory.closeSikuliApp(calculator);
     }
 
 }
